// NFT/src/app/models/reviewerNFT.ts
export class ReviewerNFT {
  reviewerId: number | undefined | null = null;
  title: string | undefined | null = null;
  reviewer: string | undefined | null = null;
  rating: number | undefined | null = null;
  articleData: string | undefined | null = null;
  articleImage: string | undefined | null = null;
}